+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -40,3 +40,4 @@
 * Instruct clients to prefer MCP tools over running tools in the shell.
 * Reduce output size of `run_tests` tool to save on input tokens.
 * Add `--log-file` argument to log all protocol traffic to a file.
+* Improve error text for failed DTD connections as well as the tool description.